+++ backend/keystone.ts
@@ -1,6 +1,9 @@
 import { config } from '@keystone-next/keystone/schema';
 import { statelessSessions } from '@keystone-next/keystone/session';
-import { createAuth } from '@opensaas/keystone-nextjs-auth';
+import {
+  createAuth,
+  nextAuthProviders as Providers,
+} from '@opensaas/keystone-nextjs-auth';
 import { KeystoneContext } from '@keystone-next/types';
 import { lists } from './schemas';
 
@@ -18,6 +21,14 @@ if (!sessionSecret) {
 
 const sessionMaxAge = 60 * 60 * 24 * 30; // 30 days
 
+export const providers = [
+  Providers.Auth0({
+    clientId: process.env.AUTH0_CLIENT_ID || 'Auth0ClientID',
+    clientSecret: process.env.AUTH0_CLIENT_SECRET || 'Auth0ClientSecret',
+    domain: process.env.AUTH0_DOMAIN || 'opensaas.au.auth0.com',
+  }),
+];
+
 const auth = createAuth({
   listKey: 'User',
   identityField: 'subjectId',
@@ -26,16 +37,6 @@ const auth = createAuth({
   userMap: { subjectId: 'id', name: 'name' },
   accountMap: {},
   profileMap: { email: 'email' },
-  providers: [
-    {
-      name: 'Auth0',
-      config: {
-        clientId: process.env.AUTH0_CLIENT_ID || 'Auth0ClientID',
-        clientSecret: process.env.AUTH0_CLIENT_SECRET || 'Auth0ClientSecret',
-        domain: process.env.AUTH0_DOMAIN || 'opensaas.au.auth0.com',
-      },
-    },
-  ],
 });
 
 export default auth.withAuth(