+++ packages/keystone-nextjs-auth/src/templates/auth.ts
@@ -1,10 +1,11 @@
 import ejs from 'ejs';
-import { AuthGqlNames, Provider } from '../types';
+import { AuthGqlNames } from '../types';
 
 const template = `
 import getNextAuthPage from '@opensaas/keystone-nextjs-auth/pages/NextAuthPage';
 import { nextAuthProviders as Providers } from '@opensaas/keystone-nextjs-auth';
 import { lists } from '.keystone/api';
+import { providers }from '../../../../../keystone';
 
 export default getNextAuthPage({
         identityField: '<%= identityField %>',
@@ -15,15 +16,7 @@ export default getNextAuthPage({
         accountMap: <%- JSON.stringify(accountMap) %>,
         profileMap: <%- JSON.stringify(profileMap) %>,
         autoCreate: <%= autoCreate %>,
-        providers:[ <% for (const i in providers){ %>
-            Providers.<%= providers[i].name %>({
-                <% const providerConf = providers[i].config;
-                for (const key in providerConf) { %>
-                    <%= key %>: '<%= providerConf[key] %>',
-                <%}%>
-                }),
-        <% } %>
-        ],
+        providers,
         lists,
     });
   `;
@@ -31,7 +24,6 @@ export default getNextAuthPage({
 export const authTemplate = ({
   gqlNames,
   identityField,
-  providers,
   sessionData,
   listKey,
   autoCreate,
@@ -41,7 +33,6 @@ export const authTemplate = ({
 }: {
   gqlNames: AuthGqlNames;
   identityField: string;
-  providers: Provider[];
   sessionData: any;
   listKey: string;
   autoCreate: boolean;
@@ -49,7 +40,6 @@ export const authTemplate = ({
   const authOut = ejs.render(template, {
     gqlNames,
     identityField,
-    providers,
     sessionData,
     listKey,
     autoCreate,