+++ packages/keystone-nextjs-auth/src/types.ts
@@ -22,10 +22,6 @@ export type AuthTokenTypeConfig = {
   /** How long do tokens stay valid for from time of issue, in minutes * */
   tokensValidForMins?: number;
 };
-export type Provider = {
-  name: string;
-  config: {};
-};
 
 export type AuthConfig<GeneratedListTypes extends BaseGeneratedListTypes> = {
   /** The key of the list to authenticate users with */
@@ -34,8 +30,6 @@ export type AuthConfig<GeneratedListTypes extends BaseGeneratedListTypes> = {
   identityField: GeneratedListTypes['fields'];
   /** Session data population */
   sessionData?: string;
-  /** NextAuth Providers */
-  providers: Provider[];
   /** Auth Create users in Keystone DB from Auth Provider */
   autoCreate: boolean;
   /** Map User in next-auth to item */